requirements.txt
aenum==3.1.15
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
asttokens==2.4.1
asyncer==0.0.2
attrs==23.2.0
azure-core==1.30.1
azure-search-documents==11.0.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.7.0
blis==0.7.11
build==1.1.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.502
chardet==5.2.0
charset-normalizer==3.3.2
chevron==0.14.0
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.1
confection==0.1.4
crypto==1.4.1
cryptography==42.0.5
cymem==2.0.8
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
distro==1.9.0
emoji==2.11.1
executing==2.0.1
fastapi==0.110.1
fastapi-socketio==0.0.10
filelock==3.13.1
filetype==1.2.0
Flask==3.0.3
flatbuffers==23.5.26
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
ginza==5.1.3
gitdb==4.0.11
GitPython==3.1.42
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.62.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.2
ipykernel==6.29.3
ipython==8.22.2
isodate==0.6.1
itsdangerous==2.2.0
ja-ginza==5.1.3
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kubernetes==29.0.0
langchain==0.1.11
langchain-community==0.0.25
langchain-core==0.1.29
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langcodes==3.3.0
langdetect==1.0.9
langsmith==0.1.22
Lazify==0.4.0
line-bot-sdk==3.11.0
literalai==0.0.500
lxml==5.2.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
Naked==0.1.32
nest-asyncio==1.6.0
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
ordered-set==4.1.0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
pke @ git+https://github.com/boudinfl/pke.git@69871ffdb720b83df23684fea53ec8776fd87e63
plac==1.4.3
platformdirs==4.2.0
posthog==3.5.0
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pulsar-client==3.4.0
pure-eval==0.2.2
pyarrow==15.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.6.3
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyJWT==2.8.0
pyOpenSSL==24.1.0
pypdf==4.2.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-iso639==2024.2.7
python-magic==0.4.27
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
pyvmomi==8.0.2.0.1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
qrcode==7.4.2
rapidfuzz==3.8.1
redis==5.0.4
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentence-transformers==2.5.1
shellescape==3.8.1
simple-websocket==1.0.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.28
srsly==2.4.8
stack-data==0.6.3
starlette==0.37.2
streamlit==1.33.0
streamlit-tree-select==0.0.5
SudachiDict-core==20240109
SudachiPy==0.6.8
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.3
thinc==8.1.12
threadpoolctl==3.3.0
tiktoken==0.6.0
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
torch==2.2.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.1
transformers==4.38.2
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.3.8
unstructured==0.13.3
unstructured-client==0.22.0
uptrace==1.24.0
urllib3==2.2.1
uvicorn==0.25.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.20.0
wcwidth==0.2.13
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.3
wrapt==1.16.0
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.17.0
streamlit-ace
streamlit_code_editor==0.1.14
